Move functions to separate file and write tests
Make the tests more official and move the functions to its own file to decouple the tests
